in places where no rename is necessary, avoid the work

+++ tools/azure-sdk-tools/ci_tools/build.py
@@ -194,15 +194,16 @@ def create_package(
         # formatting a whl distributions. This patch avoids that for now.
         normalized_package_name = setup_parsed.name.replace("-", "_")
         generated_sdist_name = os.path.join(dist, f"{normalized_package_name}-{setup_parsed.version}.tar.gz")
+        destination = generated_sdist_name.replace(normalized_package_name, setup_parsed.name)
 
-        if os.path.exists(generated_sdist_name):
-            source = generated_sdist_name
-            destination = source.replace(normalized_package_name, setup_parsed.name)
+        # in cases where there is no reformatting necessary (sdk/core/corehttpp is a good example)
+        # skip this effort
+        if generated_sdist_name != destination:
+            if os.path.exists(generated_sdist_name):
+                if os.path.exists(destination):
+                    os.remove(destination)
 
-            if os.path.exists(destination):
-                os.remove(destination)
-
-            os.rename(source, destination)
-        else:
-            logger.log(level=logging.ERROR, msg=f"Unable to find a source distribution for {setup_parsed.name} that could be renamed without normalization.")
+                os.rename(generated_sdist_name, destination)
+            else:
+                logger.log(level=logging.ERROR, msg=f"Unable to find a source distribution for {setup_parsed.name} that could be renamed without normalization.")
